refactor: add more attribute - add created_at - add status - add repo, service
