+++ src/main/java/com/group2/glamping/model/entity/User.java
@@ -1,8 +1,6 @@
-package com.group2.glamping.user;
+package com.group2.glamping.model.entity;
 
-import com.group2.glamping.model.entity.Booking;
-import com.group2.glamping.model.entity.CampSite;
-import com.group2.glamping.model.entity.Report;
+import com.group2.glamping.model.entity.enums.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,6 +9,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 
@@ -23,8 +22,8 @@ import java.util.List;
 public class User implements UserDetails {
 
     @Id
-    @GeneratedValue
-    @Column(name="id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
     private int id;
 
     @Column(name = "email")
@@ -48,6 +47,12 @@ public class User implements UserDetails {
     @Enumerated(EnumType.STRING)
     @Column(name = "role")
     private Role role;
+
+    @Column(name = "created_at")
+    private LocalDateTime created_at;
+
+    @Column(name = "status")
+    private boolean status;
 
     @OneToMany(mappedBy = "user")
     private List<Booking> bookingList;
