+++ src/main/java/com/group2/glamping/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.group2.glamping.user;
+package com.group2.glamping.repository;
 
+import com.group2.glamping.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
