+++ src/main/java/com/group2/glamping/service/impl/UserService.java
@@ -1,0 +1,11 @@
+package com.group2.glamping.service.impl;
+
+import com.group2.glamping.service.interfaces.IUserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class UserService implements IUserService {
+
+}
